move some function in lib/function rename lib/index to issueCommentHandler
