+++ lib/github.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const GitHubApi = require('github');
 
 const github = new GitHubApi({
