+++ lib/issueCommentHandler.js
@@ -1,45 +1,15 @@
+'use strict';
+
 const github = require('./github');
 const bumpVersion = require('semver-increment');
 const promisify = require('es6-promisify');
+const getMasksFromCommentBody = require('./function').getMasksFromCommentBody;
+const getPrMsgFormUrl = require('./function').getPrMsgFormUrl;
+const getBranchMsgFormHtmlUrl = require('./function').getBranchMsgFormHtmlUrl;
 
 const getPullRequests = promisify(github.pullRequests.get.bind(github.pullRequests));
 const getContent = promisify(github.repos.getContent.bind(github.repos));
 const updateFile = promisify(github.repos.updateFile.bind(github.repos));
-
-function getMasksFromCommentBody(commentBody) {
-  let MAJOR = 0;
-  let MINOR = 0;
-  let PATCH = 0;
-  if (commentBody.match('major')) {
-    MAJOR = 1;
-  } else if (commentBody.match('minor')) {
-    MINOR = 1;
-  } else if (commentBody.match('patch')) {
-    PATCH = 1;
-  } else {
-    return null;
-  }
-  const masks = [MAJOR, MINOR, PATCH];
-  return masks;
-}
-
-function getMsgFormPrUrl(url) {
-  const list = url.split('/');
-  return {
-    user: list[4],
-    repo: list[5],
-    number: list[7],
-  };
-}
-
-function getMsgFormHtmlUrl(url) {
-  const list = url.split('/');
-  return {
-    user: list[3],
-    repo: list[4],
-    branch: list[6],
-  };
-}
 
 function issueCommentHandler(event) {
   const payload = event.payload;
@@ -54,7 +24,7 @@ function issueCommentHandler(event) {
     console.log(commentBody, 'It\'s not a bump action.');
     return;
   }
-  getPullRequests(getMsgFormPrUrl(pullRequests.url))
+  getPullRequests(getPrMsgFormUrl(pullRequests.url))
   .then((res) => {
     const baseMsg = {
       user: res.base.repo.owner.login,
@@ -76,7 +46,7 @@ function issueCommentHandler(event) {
     const basePackageJson = JSON.parse(new Buffer(baseRes.content, 'base64').toString());
     const headPackageJson = JSON.parse(new Buffer(headRes.content, 'base64').toString());
     headPackageJson.version = bumpVersion(masks, basePackageJson.version);
-    const msg = getMsgFormHtmlUrl(headRes.html_url);
+    const msg = getBranchMsgFormHtmlUrl(headRes.html_url);
     msg.path = 'package.json';
     msg.message = `version: ${basePackageJson.version} bump to :version ${headPackageJson.version}`;
     msg.content = new Buffer(JSON.stringify(headPackageJson, null, '  ')).toString('base64');
@@ -87,7 +57,4 @@ function issueCommentHandler(event) {
   .catch((err) => console.error(err.message && err.stack));
 }
 
-exports.getMasksFromCommentBody = getMasksFromCommentBody;
-exports.getMsgFormPrUrl = getMsgFormPrUrl;
-exports.getMsgFormHtmlUrl = getMsgFormHtmlUrl;
-exports.issueCommentHandler = issueCommentHandler;
+module.exports = issueCommentHandler;
